+++ nakadion/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BatchLine::received_at` is now the timestamp when the batch was completely received.
 - Default of `StreamDeadPolicy` is to cancel the stream after 300s with no frames received. Previous was to never abort
 
+### ADDED
+
+- log why a stream was aborted
+
 ## [0.22.5] - 2020-05-06
 
 ### CHANGED