+++ nakadion/CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [X]
+## [0.22.7] - 2020-05-08
+
+### ADDED
+
+- metrics for IO and parse errors on the stream
+
+## [0.22.6] - 2020-05-07
 
 ### CHANGED